requirements.txt
torch
numpy
tqdm
pyyaml
swanlab
gymnasium
generals-bots
